refactor: 로비 관련 도메인 패키지 변경
